# lib/api-zod/src/generated/types/tmdbDetails.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ENYGMA Streaming API
 * OpenAPI spec version: 0.1.0
 */
import type { TmdbCast } from './tmdbCast';
import type { TmdbNetwork } from './tmdbNetwork';
import type { TmdbRecommendation } from './tmdbRecommendation';
import type { TmdbVideo } from './tmdbVideo';

export interface TmdbDetails {
  id: number;
  title: string;
  /** @nullable */
  overview?: string | null;
  /** @nullable */
  posterPath?: string | null;
  /** @nullable */
  backdropPath?: string | null;
  /** @nullable */
  logoPath?: string | null;
  /** @nullable */
  releaseDate?: string | null;
  /** @nullable */
  voteAverage?: number | null;
  /** @nullable */
  voteCount?: number | null;
  genres?: string[];
  /** @nullable */
  runtime?: number | null;
  /** @nullable */
  director?: string | null;
  writers?: string[];
  spokenLanguages?: string[];
  /** @nullable */
  originalLanguage?: string | null;
  /** @nullable */
  tagline?: string | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  budget?: number | null;
  /** @nullable */
  revenue?: number | null;
  productionCountries?: string[];
  /** @nullable */
  imdbId?: string | null;
  /** @nullable */
  contentRating?: string | null;
  /** @nullable */
  numberOfSeasons?: number | null;
  /** @nullable */
  numberOfEpisodes?: number | null;
  createdBy?: string[];
  networks?: TmdbNetwork[];
  cast?: TmdbCast[];
  videos?: TmdbVideo[];
  recommendations?: TmdbRecommendation[];
}